Limit to 65K lines per file, and 65K chars per line

+++ src/parse.rs
@@ -9,8 +9,8 @@ use crate::{
 
 struct Tokenizer<'a> {
     chars: Peekable<Chars<'a>>,
-    line: usize,
-    pos: usize,
+    line: u16,
+    pos: u16,
 }
 
 #[derive(PartialEq, Debug)]
